Added mapId support. still needs many-to-many implementation fi

+++ lib/data_manipulation/delete/one_to_many/service.dart
@@ -4,13 +4,16 @@ import 'package:dart_store/data_manipulation/delete/service.dart';
 
 class OneToManyDeleteService {
   Future<void> preDelete(EntityInstance entityInstance) async {
-    for (final foreignColumnInstance in entityInstance.oneToManyColumnsInstances()) {
+    for (final foreignColumnInstance
+        in entityInstance.oneToManyColumnsInstances()) {
       final List<dynamic> values = foreignColumnInstance.value;
 
       for (final oneOfManyItems in values) {
         final EntityInstance oneOfManyItemsInstance = EntityInstanceService()
             .entityInstanceByValueInstance(oneOfManyItems);
-        await DeleteService().delete(oneOfManyItemsInstance);
+        if (!foreignColumnInstance.mapId) {
+          await DeleteService().delete(oneOfManyItemsInstance);
+        }
       }
     }
   }
